ashell: move comms-shell functionality to shell-state
