+++ src/ashell/comms-shell.c
@@ -1,394 +1,2 @@
 // Copyright (c) 2016-2017, Intel Corporation.
 
-/**
- * @file
- * @brief Shell to keep the different states of the machine
- */
-
-// C includes
-#include <stdbool.h>
-#include <stdio.h>
-#include <string.h>
-
-// Zephyr includes
-#include <atomic.h>
-#include <ctype.h>
-#include <fs.h>
-#include <malloc.h>
-#include <misc/printk.h>
-#include <zephyr/types.h>
-
-// ZJS includes
-#include "comms-shell.h"
-#include "comms-uart.h"
-#include "shell-state.h"
-
-#include "ihex-handler.h"
-
-#include "../zjs_util.h"
-
-static const char comms_default_prompt[] =
-    ANSI_FG_YELLOW "acm> " ANSI_FG_RESTORE;
-static const char *comms_prompt = NULL;
-
-void comms_set_prompt(const char *prompt)
-{
-    comms_prompt = prompt;
-}
-
-const char *comms_get_prompt()
-{
-    if (comms_prompt == NULL)
-        return comms_default_prompt;
-    return comms_prompt;
-}
-
-#ifndef CONFIG_SHELL_UPLOADER_DEBUG
-#define DBG(...) { ; }
-#else
-#define DBG printk
-#endif /* CONFIG_SHELL_UPLOADER_DEBUG */
-
-#define CMD_ECHO_OFF "echo off\n"
-
-/* ========================================================================= */
-
-void comms_print(const char *buf)
-{
-    terminal->send(buf, strnlen(buf, MAX_LINE));
-}
-
-/**
-* Provide console message implementation for the engine.
-*/
-void comms_printf(const char *format, ...)
-{
-    va_list args;
-    va_start(args, format);
-    vfprintf(stdout, format, args);
-    va_end(args);
-}
-
-/* ========================================================================= */
-
-static bool terminal_done = false;
-static bool echo_mode = true;
-
-static inline void cursor_forward(unsigned int count)
-{
-    for (int t = 0; t < count; t++)
-        comms_print("\x1b[1C");
-}
-
-static inline void cursor_backward(unsigned int count)
-{
-    for (int t = 0; t < count; t++)
-        comms_print("\x1b[1D");
-}
-
-static inline void cursor_save(void)
-{
-    comms_print("\x1b[s");
-}
-
-static inline void cursor_restore(void)
-{
-    comms_print("\x1b[u");
-}
-
-static void insert_char(char *pos, char c, u8_t end)
-{
-    char tmp;
-
-    if (end == 0) {
-        *pos = c;
-        return;
-    }
-
-    tmp = *pos;
-    *(pos++) = c;
-
-    cursor_save();
-
-    while (end-- > 0) {
-        terminal->send(&tmp, 1);
-        c = *pos;
-        *(pos++) = tmp;
-        tmp = c;
-    }
-
-    /* Move cursor back to right place */
-    cursor_restore();
-}
-
-static void del_char(char *pos, u8_t end)
-{
-    terminal->send("\b", 1);
-
-    if (end == 0) {
-        terminal->send(" ", 1);
-        terminal->send("\b", 1);
-        return;
-    }
-
-    cursor_save();
-
-    while (end-- > 0) {
-        *pos = *(pos + 1);
-        terminal->send((pos++), 1);
-    }
-
-    terminal->send(" ", 1);
-
-    /* Move cursor back to right place */
-    cursor_restore();
-}
-
-enum
-{
-    ESC_ESC,
-    ESC_ANSI,
-    ESC_ANSI_FIRST,
-    ESC_ANSI_VAL,
-    ESC_ANSI_VAL_2
-};
-
-static atomic_t esc_state;
-static unsigned int ansi_val, ansi_val_2;
-static u8_t cur, end;
-
-static void handle_ansi(u8_t byte)
-{
-    if (atomic_test_and_clear_bit(&esc_state, ESC_ANSI_FIRST)) {
-        if (!isdigit(byte)) {
-            ansi_val = 1;
-            goto ansi_cmd;
-        }
-
-        atomic_set_bit(&esc_state, ESC_ANSI_VAL);
-        ansi_val = byte - '0';
-        ansi_val_2 = 0;
-        return;
-    }
-
-    if (atomic_test_bit(&esc_state, ESC_ANSI_VAL)) {
-        if (isdigit(byte)) {
-            if (atomic_test_bit(&esc_state, ESC_ANSI_VAL_2)) {
-                ansi_val_2 *= 10;
-                ansi_val_2 += byte - '0';
-            } else {
-                ansi_val *= 10;
-                ansi_val += byte - '0';
-            }
-            return;
-        }
-
-        /* Multi value sequence, e.g. Esc[Line;ColumnH */
-        if (byte == ';' &&
-            !atomic_test_and_set_bit(&esc_state, ESC_ANSI_VAL_2)) {
-            return;
-        }
-
-        atomic_clear_bit(&esc_state, ESC_ANSI_VAL);
-        atomic_clear_bit(&esc_state, ESC_ANSI_VAL_2);
-    }
-
-ansi_cmd:
-    switch (byte) {
-    case ANSI_BACKWARD:
-        if (ansi_val > cur) {
-            break;
-        }
-
-        end += ansi_val;
-        cur -= ansi_val;
-        cursor_backward(ansi_val);
-        break;
-    case ANSI_FORWARD:
-        if (ansi_val > end) {
-            break;
-        }
-
-        end -= ansi_val;
-        cur += ansi_val;
-        cursor_forward(ansi_val);
-        break;
-    default:
-        break;
-    }
-
-    atomic_clear_bit(&esc_state, ESC_ANSI);
-}
-
-u32_t terminal_init()
-{
-    DBG("[SHELL] Init\n");
-    return 0;
-}
-
-void terminal_error(u32_t error)
-{
-}
-
-u32_t terminal_process(const char *buf, u32_t len)
-{
-    static char *shell_line = NULL;
-    static u8_t tail = 0;
-    u32_t processed = 0;
-    // printed Is used to make sure we don't re-print characters
-    u8_t printed = cur;
-    bool flush_line = false;
-    if (shell_line == NULL) {
-        DBG("[Process]%d\n", (int)len);
-        DBG("[%s]\n", buf);
-        shell_line = (char *)zjs_malloc(MAX_LINE);
-        memset(shell_line, 0, MAX_LINE);
-        tail = 0;
-    }
-
-    /* Don't send back the 'echo off' command */
-    if (strequal(CMD_ECHO_OFF, buf)) {
-        echo_mode = false;
-    }
-
-    while (len-- > 0) {
-        processed++;
-        u8_t byte = *buf++;
-
-        if (tail == MAX_LINE) {
-            DBG("Line size exceeded \n");
-            tail = 0;
-        }
-
-        /* Handle ANSI escape mode */
-        if (atomic_test_bit(&esc_state, ESC_ANSI)) {
-            handle_ansi(byte);
-            continue;
-        }
-
-        /* Handle escape mode */
-        if (atomic_test_and_clear_bit(&esc_state, ESC_ESC)) {
-            switch (byte) {
-            case ANSI_ESC:
-                atomic_set_bit(&esc_state, ESC_ANSI);
-                atomic_set_bit(&esc_state, ESC_ANSI_FIRST);
-                break;
-            default:
-                break;
-            }
-            continue;
-        }
-
-        /* Handle special control characters */
-        if (!isprint(byte)) {
-            switch (byte) {
-            case ASCII_DEL:
-            case ASCII_BKSP:
-                if (cur > 0) {
-                    del_char(&shell_line[--cur], end);
-                }
-                break;
-            case ASCII_ESC:
-                atomic_set_bit(&esc_state, ESC_ESC);
-                break;
-            case ASCII_CR:
-                DBG("<CR>\n");
-                flush_line = true;
-                break;
-            case ASCII_TAB:
-                terminal->send("\t", 1);
-                break;
-            case ASCII_IF:
-                flush_line = true;
-                DBG("<IF>");
-                break;
-            default:
-                if (echo_mode) {
-                    printf("<CTRL> %u\n", byte);
-                }
-                flush_line = true;
-                shell_line[cur++] = byte;
-                break;
-            }
-        }
-        // We have a newline, flush the current line
-        if (flush_line) {
-            DBG("Line %u %u \n", cur, end);
-            shell_line[cur + end] = '\0';
-            if (comms_get_echo_mode()) {
-                terminal->send(shell_line + printed, cur - printed);
-                terminal->send("\r\n", 2);
-            }
-
-            u32_t length = strnlen(shell_line, MAX_LINE);
-
-            ashell_main_state(shell_line, length);
-
-            printk("\n%s", system_get_prompt());
-            comms_print("\r\n");
-            comms_print(comms_get_prompt());
-
-            cur = end = printed = 0;
-            flush_line = false;
-            if (terminal_done) {
-                break;
-            }
-        } else if (isprint(byte)) {
-            /* Ignore characters if there's no more buffer space */
-            if (cur + end < MAX_LINE - 1) {
-                insert_char(&shell_line[cur++], byte, end);
-            } else {
-                DBG("Max line\n");
-            }
-        }
-    }
-    // Once the data has been parsed, print it
-    if (comms_get_echo_mode()) {
-        terminal->send(shell_line + printed, cur - printed);
-    }
-
-    /* Done processing line */
-    if (cur == 0 && end == 0 && shell_line != NULL) {
-        DBG("[Free]\n");
-        zjs_free(shell_line);
-        shell_line = NULL;
-    }
-    return processed;
-}
-
-bool terminal_process_done()
-{
-    return terminal_done;
-}
-
-void comms_set_echo_mode(bool mode)
-{
-    echo_mode = mode;
-}
-
-bool comms_get_echo_mode()
-{
-    return echo_mode;
-}
-
-u32_t terminal_close()
-{
-    DBG("[SHELL CLOSE]\n");
-    ihex_process_start();
-    return 0;
-}
-
-static struct terminal_config terminal_cfg = {
-    .init = terminal_init,
-    .error = terminal_error,
-    .done = terminal_process_done,
-    .close = terminal_close,
-    .process = terminal_process,
-    .send = uart_write_buf
-};
-
-void terminal_start()
-{
-    terminal_done = false;
-    terminal = &terminal_cfg;
-}
